requirements.txt
dnspython==2.6.1
pymongo==4.8.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dnspython==2.6.1
email_validator==2.2.0
eventlet==0.36.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-SocketIO==5.3.6
folium==0.17.0
fonttools==4.53.0
frozenlist==1.4.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
iso8601==2.1.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
libusb1==3.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pynmea2==1.19.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
python-usbtmc==0.8
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
rich==13.7.1
serial==0.0.97
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
usb==0.0.83.dev0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
xyzservices==2024.6.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
